Did many changes and more explanation in the comments. It's well documented
