+++ docs/test_doc.md
@@ -1,0 +1,8 @@
+# Big Heading Test
+- list
+
+$$
+(A - \lambda I_n)\vec{x} = 0 \\ - eq 1
+$$
+
+See if the above equation appears in Sphinx
