+++ robotic_skin/sensor/__init__.py
@@ -1,0 +1,1 @@
+from .sensor import Sensor
