+++ robotic_skin/sensor/adxl335.py
@@ -1,80 +1,52 @@
 import math
 import time
 import numpy as np
+from robotic_skin.sensor import Sensor
 from mcp3208 import MCP3208
 
-ADCONVERTER_BIT = 12
-MIN_DISCRETE_SIGNAL = 0
-MAX_DISCRETE_SIGNAL = math.pow(2, ADCONVERTER_BIT)
-MAX_G = 3
-VOLTAGE_SCALE  = 5/3.3
-
-class Sensor(object):
-    """
-    Sensor Class
-    """
-    def __init__(self):
-        pass
-
-    def calibrate(self):
-        """
-        Calibrate the sensor
-        """
-        raise NotImplementedError()
-
-    def read(self):
-        """
-        Fetch sensor data
-        """
-        raise NotImplementedError()
 
 class ADXL335(Sensor):
     def __init__(self, xpin, ypin, zpin):
         super(ADXL335, self).__init__()
+        # We are using MPC3208 library from Pypi, thank God someone wrote it
         self.adc = MCP3208()
         self.pins = [xpin, ypin, zpin]
+        # self.data is holder for all our accelerations in x,y,z directions
         self.data = np.zeros(len(self.pins))
-        self.calibrated = False
-        self.mins = np.inf*np.ones(3)
-        self.maxs = -np.inf*np.ones(3)
-        self.bias = (MAX_DISCRETE_SIGNAL-MIN_DISCRETE_SIGNAL) / 2
-        self.scale = (2*MAX_G) / (MAX_DISCRETE_SIGNAL-MIN_DISCRETE_SIGNAL)
+        self.calibrate = False
         print('Make sure to supply 5V to the circuit & 3.3V to AD Converter Vref')
 
-    def calibrate(self, n_loops=10000):
-        """
-        Depreciated
-        """
-        print('start calibrating')
-        for i in range(n_loops):
-            data = self.read()
-            for j in range(3):
-                self.mins[j] = min(self.mins[j], data[j])
-                self.maxs[j] = max(self.maxs[j], data[j])
-            print(self.mins, self.maxs)
-        
-        self.bias = (self.mins + self.maxs) / 2
-        self.scale = 3 / (self.maxs-self.bias)
-        self.calibrated = True
+    def calibrate(self) -> None:
+        # How dumb can it be to write some shit like this that calibrate = True in calibrate function?
+        # Well we gotta calibrate it before. It isn't automated, which I hate, but I gotta do what I gotta do
+        # Else Satan help me I will forget if I don't write this
+        # This is basically implemented to satisfy the Sensor class, as I have to override the class for sure
+        self.calibrate = True
 
     def _read_raw(self):
         """
-        returns python list of fetched sensor data
+        returns python list of fetched sensor data. The data ranges from [0, 4096]
+        It's basically an ADC output. As MPC3208 is a 12 bit ADC, 2^12 = 4096, hence the 4096 is the
+        upper limit
         """
         return [self.adc.read(pin) for pin in self.pins]
 
     def read(self):
-        """
-        returns numpy list of fetched sensor data
-        """
-        data = VOLTAGE_SCALE*np.array(self._read_raw())
-        self.data = self.scale * (data - self.bias)
+        # Read raw ADC values from the sensor
+        data = np.array(self._read_raw())
+        # How this equation came up, look for my blog:
+        # https://krishnachaitanya9.github.io/posts/RPi_Calibrating_ADXL335_Accelerometer/
+        # There is a more detailed explanation
+        # which can't be fir into comments. So please go look, and lemme know if there is any problem with the logic
+        self.data = ((((3.3 / 4096) * data) - 2) * 3.1) + 1
+        # Returning back the data
         return self.data
+
 
 if __name__ == '__main__':
     accel_sensor = ADXL335(xpin=0, ypin=1, zpin=2)
-
+    accel_sensor.calibrate
     while True:
         data = accel_sensor.read()
-        print(data[0], data[1], data[2])
+        print("Z Acceleration: ", data[2])
         time.sleep(0.5)
